server/requirements.txt
alembic==1.6.5
bcrypt==3.2.0
Bcrypt-Flask==1.0.1
cffi==1.14.5
click==8.0.1
Flask==2.0.1
Flask-JWT-Extended==4.2.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.19
Werkzeug==2.0.1
